+++ src/constants.ts
@@ -5,6 +5,7 @@ export const Steps = {
   USERS: 'step-users',
   GROUPS: 'step-groups',
   GROUP_MEMBERS: 'step-group-members',
+  INCIDENTS: 'step-incidents',
 };
 
 export const Entities = {
@@ -22,6 +23,11 @@ export const Entities = {
     resourceName: 'User Group',
     _type: 'service_now_group',
     _class: 'UserGroup',
+  },
+  INCIDENT: {
+    resourceName: 'Incident',
+    _type: 'service_now_incident',
+    _class: 'Incident',
   },
 };
 
@@ -49,5 +55,11 @@ export const Relationships = {
     sourceType: Entities.GROUP._type,
     _class: RelationshipClass.HAS,
     targetType: Entities.GROUP._type,
+  },
+  INCIDENT_ASSIGNED_USER: {
+    _type: 'service_now_incident_assigned_user',
+    sourceType: Entities.INCIDENT._type,
+    _class: RelationshipClass.ASSIGNED,
+    targetType: Entities.USER._type,
   },
 };